requirements.txt
dj-database-url==0.4.1
psycopg2==2.6.2
Django==1.11.5
djangorestframework==3.7.3
gunicorn==19.6.0

django-db-file-storage==0.4.5